factory and seed

+++ larazillow11/database/migrations/2024_11_10_012605_add_fields_to_listings_table.php
@@ -14,12 +14,12 @@ return new class extends Migration
         Schema::table('listings', function (Blueprint $table) {
             $table->unsignedTinyInteger('beds');
             $table->unsignedTinyInteger('baths');
-            $table->unsignedTinyInteger('area');
+            $table->smallInteger('area');
             $table->tinyText('city');
             $table->tinyText('code');
             $table->tinyText('street');
             $table->tinyText('street_nr');
-            $table->unsignedTinyInteger('price');
+            $table->unsignedInteger('price');
         });
     }
